More explicit & robust windows services-related scripts. (#1338)
This PR replaces relative paths to executables and files with full paths using %~dp0 placeholder that denotes directory with the script. Thanks to that scripts work correctly even when run from different working directory.
Quotes are necessary to properly handle spaces in paths.
The /y flag on copy is not necessary (it is implied by default when executing scripts), however I added it, so the code behaves the same way as when executed directly in the command prompt.

Note: the arguments to sc.exe like s-luna-empire are not files but sercive identifiers, and as such are not to be modified here.