Allow copy as fallback when creating symlink fails (#1340)
See the in-code comment and issue https://github.com/luna/luna-manager/issues/236
This PR allows packaging on Windows without administrator privileges. If creating symlink fails, the file will be just copied. It is not nice workaround, but symlinks are replaced with copies on Windows anyway later in process. Now it is at least described.