Don't rely on PWD environemnt variable. Make supervisord call cwd-independent. ref #1385 (#1386)
This PR fixes #1385 — a regression introduced when moving away from shelly. As I recommended a shell call was replaced with direct process spawn. However, the spawned `supervisord` not only relies on its working directory — it also needed a PWD environment variable, as it was used in our configuration files.

This PR fixes this:
1) by using relative paths in configuration files — so they work no matter what the PWD value is (or when it is not present at all)
2) the runner shall pass absolute path to the configuration file to the supervisord — so it can be spawned without adjusting working directory